Rename logger file to utils in order to use it to store more utils functions
