+++ deployability/modules/allocation/aws/credentials.py
@@ -5,7 +5,7 @@ from botocore.exceptions import ClientError
 from pathlib import Path
 
 from modules.allocation.generic import Credentials
-from modules.allocation.generic.logger import logger
+from modules.allocation.generic.utils import logger
 
 
 class AWSCredentials(Credentials):
