+++ deployability/modules/allocation/vagrant/provider.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 from modules.allocation.generic import Provider
 from modules.allocation.generic.models import CreationPayload
-from modules.allocation.generic.logger import logger
+from modules.allocation.generic.utils import logger
 from .credentials import VagrantCredentials
 from .instance import VagrantInstance
 from .models import VagrantConfig
