+++ deployability/modules/provision/actions.py
@@ -2,7 +2,7 @@ from modules.generic import Ansible
 
 from modules.provision.componentType import Package, AIO, Generic, Dependencies
 from modules.provision.models import ComponentInfo
-from modules.provision.logger import logger
+from modules.provision.utils import logger
 
 class Action:
     """
