+++ deployability/modules/workflow_engine/utils.py
@@ -1,0 +1,3 @@
+from modules.generic.logger import Logger
+
+logger = Logger("workflow_enine")
