+++ deployability/modules/workflow_engine/workflow_processor.py
@@ -2,16 +2,17 @@
 # Created by Wazuh, Inc. <info@wazuh.com>.
 # This program is a free software; you can redistribute it and/or modify it under the terms of GPLv2
 
-import graphlib
 import concurrent.futures
-import time
+import graphlib
 import json
-import logging
-from itertools import product
+import time
 import yaml
-from .task import *
 
-logger = (lambda: logging.getLogger())()
+from itertools import product
+
+from .task import *
+from .utils import logger
+
 
 class WorkflowFile:
     """Class for loading and processing a workflow file."""
